Add calibration image display code
For some reason findCirclesGrid always returns false with the current set of images.
I am unsure of what the reason is at this point. Also, due to the size of the images, the current process takes some time.

+++ VideoSLAM/MarkerInfo.h
@@ -43,7 +43,7 @@ public:
     */
     cv::Mat current_camera_pose_position;
     // Mat current_camera_pose_orientation;
-    cv::Quat<float> current_camera_pose_orientation;
+    cv::Quat<double> current_camera_pose_orientation;
